+++ migrations/2019-08-17_add_event_image_urls/up.sql
@@ -1,0 +1,3 @@
+-- Add image and image link URLs to events
+ALTER TABLE events ADD COLUMN image_url TEXT;
+ALTER TABLE events ADD COLUMN image_link_url TEXT;